Backend/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
pydantic>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Testing dependencies
pytest>=7.4.3
httpx>=0.25.1  # For TestClient
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
